My solution to the Interview-bot challenge
I chose to leverage libraries when solving the different types of
exam questions in order to avoid "re-inventing the wheel" for this
challenge and produce my solution as quickly as possible.

My solution could be improved upon by optimizing my function for
solving string equations by implementing a mapping function to
reduce the amount of loops required, however I'm satisfied with the
current performance, as it completes the exam in ~1.5 seconds total.